// src/app/api/model/sparePartUsage.ts
/**
 * Service Support API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SparePart } from './sparePart';


export interface SparePartUsage { 
    sparePart?: SparePart;
    quantity?: number;
}