// src/app/api/model/worksheetDetail.ts
/**
 * Service Support API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WorksheetStatus } from './worksheetStatus';
import { WorksheetNote } from './worksheetNote';
import { Defect } from './defect';
import { UserSummary } from './userSummary';
import { SparePartUsage } from './sparePartUsage';
import { Tool } from './tool';


export interface WorksheetDetail { 
    worksheetId?: number;
    status?: WorksheetStatus;
    isWarranty?: boolean;
    hasWarrantyCard?: boolean;
    hasInvoiceCopy?: boolean;
    hasRegistrationProof?: boolean;
    createdAt?: string;
    ownerDescription?: string;
    tool?: Tool;
    defects?: Array<Defect>;
    notes?: Array<WorksheetNote>;
    spareParts?: Array<SparePartUsage>;
    assignedUser?: UserSummary;
}
export namespace WorksheetDetail {
}